Started conversion to Maven 2.
git-svn-id: https://svn.apache.org/repos/asf/webservices/xmlrpc/branches/b20050512_streaming@331034 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/xmlrpc/webserver/WebServer.java
@@ -52,7 +52,7 @@ public class WebServer implements Runnable {
 					if ("*".equals(next)) {
 						pattern[i] = 256;
 					} else {
-						pattern[i] = (byte) Integer.parseInt(next);
+						pattern[i] = Integer.parseInt(next);
 					}
 				}
 			} catch (Exception e) {
